refactor(document): 统一操作人ID数据类型并优化文档生命周期逻辑
- 将API中的OperatorId字段从string统一修改为int64类型
- 重命名DocumentLifecycleLogicImpl为LifecycleLogicImpl
- 添加文档转换器实现文档请求响应转换功能
- 新增文档异步处理业务逻辑接口定义
- 更新依赖注入绑定配置适配类名变更
- 在服务层实现文档上传转换和处理逻辑

+++ internal/domain/document/model/entity/document_strategy.go
@@ -36,7 +36,7 @@ func (d *DocumentStrategyPlan) FillEnumNames() {
 type DocumentStrategyStep struct {
 	ID                int64  `gorm:"column:id;primaryKey"`    // 主键ID
 	DocumentId        int64  `gorm:"column:document_id"`      // 文档ID
-	PlanId            int    `gorm:"column:plan_id"`          // 方案ID
+	PlanId            int64  `gorm:"column:plan_id"`          // 方案ID
 	StepNo            int    `gorm:"column:step_no"`          // 步骤序号
 	PipelineType      int    `gorm:"column:pipeline_type"`    // 流水线类型
 	StrategyType      int    `gorm:"column:strategy_type"`    // 策略类型
